requirements.txt
crewai[tools]==0.102.0
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.12.0 